Share::put was completed

+++ src/Api/Endpoint/Example.php
@@ -4,15 +4,19 @@ use Yuforms\Api\Core\Endpoint;
 
 // customable endpoint example
 class Example extends Endpoint {
-    protected function controllerHandler() {
-        $class = 'Yuforms\Api\Controller\Example';
-        new $class([
-            'method'=>'DELETE',
-            'data'=>[
-                'this is data'=>'this is data'
-            ],
-            'who'=>'admin',
-            'userId'=>3
-        ]);
+    protected function handleController() {
+        if($this->method==='POST') {
+            $class = 'Yuforms\Api\Controller\Example';
+            new $class([
+                'method'=>'DELETE',
+                'data'=>[
+                    'this is data'=>'this is data'
+                ],
+                'who'=>'admin',
+                'userId'=>3
+            ]);
+        } else {
+            parent::handleController();
+        }
     }
 }
